MAJOR (assets 🧲): Updated linux.svg file
[assets] Updated linux.svg file

The linux.svg file in the assets directory has been updated. The changes include adding an SVG element with a "viewBox" attribute and a "fill" attribute for the path element. The content of the file remains the same, but the formatting has been modified. No new lines have been added at the end of the file.