FIX (core 🧱): Update user-solid.svg file
[Fix] Update user-solid.svg file

The user-solid.svg file has been updated to fix an issue. The changes include adding a new path element to the SVG file. The viewBox attribute remains the same with values "0 0 448 512".

The previous version of the file had an incomplete opening SVG tag, which has been fixed.

The updated file no longer has a newline at the end of the file, similar to the previous version.

This commit addresses the changes described above.